Initial Html css commit
Changed the css and html file names. Added the youtube video through a figure caption. Added the audio clip for fire and ice. Added the mp3 version and the ogg version. Added the embedded audio of the file for backup. Added the audio clip for devotion. Added the mp3 version and the ogg version. Added the embedded audio of the file for backup. Called in the css style sheet to render the page with the css styles.  Made audio element render as block, gave the audio elements a dark box shadow, Gave top bottom margin 20px and left and right margin auto. Made audio have width of 80%. the color of the audio clips to blue.